Adjust numeric IDs to Long

+++ src/main/java/com/mercadopago/resources/preference/PreferenceSearch.java
@@ -69,7 +69,7 @@ public class PreferenceSearch extends MPResource {
   private String siteId;
 
   /** Sponsor id. */
-  private Integer sponsorId;
+  private Long sponsorId;
 
   /** Shipping mode. */
   private String shippingMode;
